feat: get plugin resource service details
