+++ app/src/main/java/com/kylecorry/trail_sense/plugins/domain/PluginResourceServiceFeatures.kt
@@ -1,0 +1,6 @@
+package com.kylecorry.trail_sense.plugins.domain
+
+data class PluginResourceServiceFeatures(
+    val weather: List<String>,
+    val mapLayers: List<String>
+)
